requirements.txt
click==8.0.1
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.17.1
SQLAlchemy==1.4.18
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
Flask-Migrate==3.0.1
psycopg2-binary==2.9.1
